requirements.txt
aiofiles
aiohttp
fastapi
keras
librosa
mediapipe
micloud
netifaces
numpy
opencv_contrib_python==4.7.0.68
opencv_python==4.6.0.66
opencv_python_headless==4.8.1.78
openvino==2022.3.1
openvino_dev==2022.3.1
Pillow
protobuf==3.20.3
psutil
PyAudio
pygrabber
pyserial
pyusb
PyYAML
sounddevice
tensorflow_gpu
uvicorn
websocket_client
websockets
